requirements.txt
# Astro Runtime includes the following pre-installed providers packages: https://www.astronomer.io/docs/astro/runtime-image-architecture#provider-packages
apache-airflow-providers-http
apache-airflow-providers-amazon
astronomer-cosmos
apache-airflow-providers-snowflake